+++ src/core/QueryManager.ts
@@ -104,6 +104,10 @@ import {
 } from '../data/mutationResults';
 
 import {
+  MutationStore,
+} from '../mutations/store';
+
+import {
   QueryScheduler,
 } from '../scheduler/scheduler';
 
@@ -137,6 +141,7 @@ export class QueryManager {
   public store: ApolloStore;
   public networkInterface: NetworkInterface;
   public ssrMode: boolean;
+  public mutationStore: MutationStore = new MutationStore();
 
   private addTypename: boolean;
   private deduplicator: Deduplicator;
@@ -313,6 +318,8 @@ export class QueryManager {
       update: updateWithProxyFn,
     });
 
+    this.mutationStore.initMutation(mutationId, mutationString, variables);
+
     return new Promise<ExecutionResult>((resolve, reject) => {
       this.networkInterface.query(request)
         .then((result) => {
@@ -325,6 +332,8 @@ export class QueryManager {
               error,
               mutationId,
             });
+
+            this.mutationStore.markMutationError(mutationId, error);
 
             delete this.queryDocuments[mutationId];
             reject(error);
@@ -342,6 +351,8 @@ export class QueryManager {
             updateQueries: generateUpdateQueriesInfo(),
             update: updateWithProxyFn,
           });
+
+          this.mutationStore.markMutationResult(mutationId);
 
           // If there was an error in our reducers, reject this promise!
           const { reducerError } = this.getApolloState();
@@ -836,6 +847,8 @@ export class QueryManager {
       type: 'APOLLO_STORE_RESET',
       observableQueryIds: Object.keys(this.observableQueries),
     });
+
+    this.mutationStore.reset();
 
     // Similarly, we have to have to refetch each of the queries currently being
     // observed. We refetch instead of error'ing on these since the assumption is that