+++ test/client.ts
@@ -182,7 +182,6 @@ describe('client', () => {
       {
         apollo: {
           queries: {},
-          mutations: {},
           data: {},
           optimistic: [],
           reducerError: null,
@@ -593,7 +592,6 @@ describe('client', () => {
           metadata: null,
         },
       },
-      mutations: {},
       reducerError: null,
     }) };